refactor: migrate domain models to Pydantic and restructure tests
- Replace dataclass with Pydantic BaseModel for Location and Forecast
  - Adds runtime validation and coercion as a bonus
  - Use ConfigDict(frozen=True) for immutability guarantee
- Remove TYPE_CHECKING guard in forecast.py to fix get_type_hints()
  - Runtime imports of datetime and Location now safe (no circular deps)
- Change temperatures_c from list[float] to tuple[float, ...] for true immutability
- Split multi-import statement in __init__.py (one import per line)
- Restructure unit tests under tests/unit/domain/ to mirror source tree
  - test_location.py: Location immutability tests
  - test_unit_system.py: UnitSystem enum tests
  - test_forecast.py: Forecast and public exports tests
- Add pydantic>=2.0 to project dependencies

Addresses code review feedback from both human reviewer and CodeRabbit.
