+++ pyproject.toml
@@ -30,6 +30,7 @@ classifiers = [
 ]
 dependencies = [
     "httpx>=0.27",
+    "pydantic>=2.0",
 ]
 
 [project.optional-dependencies]
