+++ tests/unit/domain/test_forecast.py
@@ -1,31 +1,10 @@
-"""Tests for domain primitives targeted for v0.1.0."""
+"""Tests for Forecast domain primitive."""
 
 from datetime import UTC, datetime
-
-import pytest
 
 import openmeteo
 from openmeteo.domain.forecast import Forecast
 from openmeteo.domain.location import Location
-from openmeteo.domain.units import UnitSystem
-
-
-def test_location_is_immutable_value_object() -> None:
-    """Location should preserve coordinates and reject mutation."""
-    location = Location(latitude=51.0504, longitude=13.7373, name="Dresden")
-
-    assert location.latitude == 51.0504
-    assert location.longitude == 13.7373
-    assert location.name == "Dresden"
-
-    with pytest.raises(AttributeError):
-        location.latitude = 0.0  # type: ignore[misc]
-
-
-def test_unit_system_values() -> None:
-    """UnitSystem should expose stable public wire values."""
-    assert UnitSystem.METRIC.value == "metric"
-    assert UnitSystem.IMPERIAL.value == "imperial"
 
 
 def test_forecast_holds_time_window_and_temperature() -> None:
@@ -37,13 +16,13 @@ def test_forecast_holds_time_window_and_temperature() -> None:
         location=Location(latitude=51.0504, longitude=13.7373, name="Dresden"),
         start=start,
         end=end,
-        temperatures_c=[3.2, 3.5],
+        temperatures_c=(3.2, 3.5),
     )
 
     assert forecast.location.name == "Dresden"
     assert forecast.start == start
     assert forecast.end == end
-    assert forecast.temperatures_c == [3.2, 3.5]
+    assert forecast.temperatures_c == (3.2, 3.5)
 
 
 def test_public_exports_include_domain_primitives() -> None:
